Mark `Spliterators` as `@AnnotatedFor("nullness")`.
Compare https://github.com/jspecify/jdk/pull/66.

One thing we could discuss here is what to do about
[`spliterator(Object[], int`](https://docs.oracle.com/en/java/javase/22/docs/api/java.base/java/util/Spliterators.html#spliterator(java.lang.Object%5B%5D,int))
and friends. In the JSpecify PR, I annotate the array-element type as
`@Nullable`. That's unsound, since the API lets you create a
`Spliterator` of a non-nullable type `T` from a `@Nullable Object[]`. In
fairness, the entire API is unsound, since it also lets you create a
`Spliterator<String>` from an `Integer[]` :) But we don't have to make
it worse. With the Checker Framework, we can use `@PolyNull`, which
seems like probably the right tool. I'm using it in this PR.